+++ src/Config/AccessLog.php
@@ -2,11 +2,12 @@
 
 namespace UnitPhpSdk\Config;
 
-use UnitPhpSdk\Builders\EndpointBuilder;
+use UnitPhpSdk\Contracts\Arrayable;
 use UnitPhpSdk\Contracts\Uploadable;
+use UnitPhpSdk\Enums\ApiPathEnum;
 use UnitPhpSdk\Traits\CanUpload;
 
-class AccessLog implements Uploadable
+class AccessLog implements Uploadable, Arrayable
 {
     use CanUpload;
 
@@ -22,15 +23,29 @@ class AccessLog implements Uploadable
      *
      * @var string|mixed|null
      */
-    private ?string $format;
+    private string|array|null $format;
+
+    /**
+     * The access_log can be dynamically turned on and off by using the if option
+     *
+     * This feature lets users set conditions to determine whether access logs are recorded.
+     * The if option supports a string and JavaScript code.
+     * If its value is empty, 0, false, null, or undefined,
+     * the logs will not be recorded. And the ‘!’ as a prefix inverses the condition.
+     *
+     * @var string|mixed|null
+     */
+    private ?string $if;
 
     public function __construct(
         array $data
-    ) {
+    )
+    {
         $this->path = $data['path'] ?? null;
         $this->format = $data['format'] ?? null;
+        $this->if = $data['if'] ?? null;
 
-        $this->setApiEndpoint(EndpointBuilder::create('/config/access_log')->get());
+        $this->setApiEndpoint(ApiPathEnum::ACCESS_LOG->value);
     }
 
     /**
@@ -50,9 +65,9 @@ class AccessLog implements Uploadable
     }
 
     /**
-     * @return string|null
+     * @return string|array|null
      */
-    public function getFormat(): ?string
+    public function getFormat(): string|array|null
     {
         return $this->format;
     }
@@ -64,4 +79,27 @@ class AccessLog implements Uploadable
     {
         $this->format = $format;
     }
+
+    /**
+     * @inheritDoc
+     */
+    #[\Override] public function toArray(): array
+    {
+        return [
+            'if' => $this->if,
+            'path' => $this->path,
+            'format' => $this->format,
+        ];
+    }
+
+    public function getIf(): ?string
+    {
+        return $this->if;
+    }
+
+    public function setIf(?string $if): void
+    {
+        $this->if = $if;
+    }
+
 }