+++ src/Config/Listener.php
@@ -8,10 +8,10 @@ use UnitPhpSdk\Config\Listener\{
     Tls
 };
 use Override;
-use UnitPhpSdk\Builders\EndpointBuilder;
 use UnitPhpSdk\Contracts\Arrayable;
 use UnitPhpSdk\Contracts\Jsonable;
 use UnitPhpSdk\Contracts\Uploadable;
+use UnitPhpSdk\Enums\ApiPathEnum;
 use UnitPhpSdk\Exceptions\UnitException;
 use UnitPhpSdk\Traits\CanUpload;
 
@@ -36,14 +36,33 @@ class Listener implements Uploadable, Arrayable, Jsonable
 
     public function __construct(
         private readonly string $listener,
+        /**
+         * Destination to which the listener passes incoming requests
+         * @var string|ListenerPass
+         */
         string|ListenerPass     $pass,
+        /**
+         * Defines SSL/TLS settings.
+         * @var Tls|null
+         */
         private ?Tls            $tls = null,
+        /**
+         * Configures client IP address and protocol replacement.
+         * @var Forwarded|null
+         */
         private ?Forwarded      $forwarded = null,
-    ) {
+        /**
+         * Controls the ‘backlog’ parameter to the listen(2) system-call.
+         * This essentially limits the number of pending connections waiting to be accepted.
+         * @var int
+         */
+        private ?int            $backlog = null
+    )
+    {
         $this->parsePort();
         $this->parsePass($pass);
 
-        $this->setApiEndpoint(EndpointBuilder::create($this)->get() . '/' . $this->getListener());
+        $this->setApiEndpoint(ApiPathEnum::LISTENER->getPath($this->getListener()));
     }
 
     /**
@@ -147,9 +166,9 @@ class Listener implements Uploadable, Arrayable, Jsonable
     /**
      * Get tls section
      *
-     * @return Tls
+     * @return Tls|null
      */
-    public function getTls(): Tls
+    public function getTls(): ?Tls
     {
         return $this->tls;
     }
@@ -253,5 +272,15 @@ class Listener implements Uploadable, Arrayable, Jsonable
     public function getTarget(): ListenerPass
     {
         return $this->getPass();
+    }
+
+    public function getBacklog(): ?int
+    {
+        return $this->backlog;
+    }
+
+    public function setBacklog(?int $backlog): void
+    {
+        $this->backlog = $backlog;
     }
 }