+++ src/Config/Route.php
@@ -2,11 +2,13 @@
 
 namespace UnitPhpSdk\Config;
 
+use Override;
 use SplObjectStorage;
 use UnitPhpSdk\Builders\EndpointBuilder;
 use UnitPhpSdk\Config\Routes\RouteBlock;
 use UnitPhpSdk\Contracts\RouteInterface;
 use UnitPhpSdk\Contracts\Uploadable;
+use UnitPhpSdk\Enums\ApiPathEnum;
 use UnitPhpSdk\Exceptions\UnitException;
 use UnitPhpSdk\Traits\CanUpload;
 use UnitPhpSdk\Traits\HasListeners;
@@ -36,9 +38,10 @@ class Route implements RouteInterface, Uploadable
      */
     public function __construct(
         private readonly string $name,
-        $data = [],
+                                $data = [],
         bool                    $single = false
-    ) {
+    )
+    {
         $this->routeBlocks = new SplObjectStorage();
         if (!empty($data)) {
             if ($single) {
@@ -50,7 +53,7 @@ class Route implements RouteInterface, Uploadable
             }
         }
 
-        $this->setApiEndpoint(EndpointBuilder::create($this)->get() . '/' . $this->getName());
+        $this->setApiEndpoint(ApiPathEnum::ROUTE->getPath($this->getName()));
     }
 
     /**
@@ -63,6 +66,7 @@ class Route implements RouteInterface, Uploadable
 
     /**
      * @param array $routeBlocks
+     * @throws UnitException
      */
     public function setRouteBlocks(array $routeBlocks): void
     {
@@ -71,6 +75,16 @@ class Route implements RouteInterface, Uploadable
                 $routeBlock instanceof RouteBlock ? $routeBlock : new RouteBlock($routeBlock)
             );
         }
+    }
+
+    public function getRouteBlock($index): RouteBlock
+    {
+        return $this->routeBlocks->offsetGet($index);
+    }
+
+    public function removeRouteBlock($index): void
+    {
+        $this->routeBlocks->offsetUnset($index);
     }
 
     /**
@@ -99,17 +113,17 @@ class Route implements RouteInterface, Uploadable
     /**
      * @return array
      */
-    #[\Override] public function toArray(): array
+    #[Override] public function toArray(): array
     {
-        return array_map(fn (RouteBlock $routeBlock) => $routeBlock->toArray(), $this->getRouteBlocks());
+        return array_map(fn(RouteBlock $routeBlock) => $routeBlock->toArray(), $this->getRouteBlocks());
     }
 
     /**
      * @param int $options
      * @return string
      */
-    #[\Override] public function toJson(int $options = 0): string
+    #[Override] public function toJson(int $options = 0): string
     {
-        return json_encode(array_filter($this->toArray(), fn ($item) => !empty($item)));
+        return json_encode(array_filter($this->toArray(), fn($item) => !empty($item)));
     }
 }