+++ src/Config/Settings.php
@@ -2,16 +2,21 @@
 
 namespace UnitPhpSdk\Config;
 
+use InvalidArgumentException;
 use UnitPhpSdk\Builders\EndpointBuilder;
 use UnitPhpSdk\Config\Settings\Http;
+use UnitPhpSdk\Config\Settings\Telemetry;
 use UnitPhpSdk\Contracts\Arrayable;
 use UnitPhpSdk\Contracts\Jsonable;
 use UnitPhpSdk\Contracts\Uploadable;
+use UnitPhpSdk\Enums\ApiPathEnum;
 use UnitPhpSdk\Traits\CanUpload;
 
 class Settings implements Uploadable, Arrayable, Jsonable
 {
     use CanUpload;
+
+    private const array REQUIRED_TELEMETRY_KEYS = ['endpoint', 'protocol'];
 
     /**
      * Fine-tunes handling of HTTP requests from the clients
@@ -25,6 +30,11 @@ class Settings implements Uploadable, Arrayable, Jsonable
      */
     private string|array $js_module = [];
 
+    /**
+     * @var Telemetry|mixed|null
+     */
+    private ?Telemetry $telemetry = null;
+
     public function __construct(array $data = [])
     {
         if (array_key_exists('http', $data)) {
@@ -35,7 +45,11 @@ class Settings implements Uploadable, Arrayable, Jsonable
             $this->parseJsModule($data['js_module']);
         }
 
-        $this->setApiEndpoint(EndpointBuilder::create($this)->get());
+        if (array_key_exists('telemetry', $data)) {
+            $this->parseTelemetry($data['telemetry']);
+        }
+
+        $this->setApiEndpoint(ApiPathEnum::SETTINGS->value);
     }
 
     /**
@@ -54,6 +68,33 @@ class Settings implements Uploadable, Arrayable, Jsonable
     private function parseJsModule(array|string $data): void
     {
         $this->js_module = $data;
+    }
+
+    private function parseTelemetry(array $data): void
+    {
+        $this->validateTelemetryData($data);
+
+        $this->telemetry = new Telemetry(
+            endpoint: $data['endpoint'],
+            protocol: $data['protocol']
+        );
+
+        if (array_key_exists('sampling_ratio', $data)) {
+            $this->telemetry->setSamplingRatio($data['sampling_ratio']);
+        }
+
+        if (array_key_exists('batch_size', $data)) {
+            $this->telemetry->setBatchSize($data['batch_size']);
+        }
+    }
+
+    private function validateTelemetryData(array $data): void
+    {
+        foreach (self::REQUIRED_TELEMETRY_KEYS as $key) {
+            if (empty($data[$key])) {
+                throw new InvalidArgumentException("Telemetry {$key} is required");
+            }
+        }
     }
 
     /**
@@ -85,6 +126,6 @@ class Settings implements Uploadable, Arrayable, Jsonable
      */
     #[\Override] public function toJson(int $options = 0): string
     {
-        return json_encode(array_filter($this->toArray(), fn ($item) => !empty($item)), $options);
+        return json_encode(array_filter($this->toArray(), fn($item) => !empty($item)), $options);
     }
 }