+++ src/Contracts/UnitInterface.php
@@ -49,10 +49,9 @@ interface UnitInterface extends Arrayable, Jsonable
      *
      * @param string $path
      * @param string $certificateName
-     * @return bool
      * @throws UnitException
      */
-    public function uploadCertificate(string $path, string $certificateName): bool;
+    public function uploadCertificate(string $path, string $certificateName);
 
     /**
      * Return full config uploaded on Nginx Unit