+++ src/Enums/TelemetryProtocolEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace UnitPhpSdk\Enums;
+
+enum TelemetryProtocolEnum: string
+{
+    case HTTP = 'http';
+    case GRPC = 'grpc';
+}