+++ src/Traits/CanUpload.php
@@ -2,6 +2,11 @@
 
 namespace UnitPhpSdk\Traits;
 
+use UnitPhpSdk\Abstract\AbstractApplication;
+use UnitPhpSdk\Config\Listener;
+use UnitPhpSdk\Config\Route;
+use UnitPhpSdk\Config\Upstream;
+use UnitPhpSdk\Enums\ApiPathEnum;
 use UnitPhpSdk\Enums\HttpMethodsEnum;
 use UnitPhpSdk\Exceptions\UnitException;
 use UnitPhpSdk\Http\UnitRequest;
@@ -48,18 +53,36 @@ trait CanUpload
     public function upload(UnitRequest $request)
     {
         $data = $this->removeEmptyArrays($this->toArray());
-
-        echo '<pre>';
-        print_r(json_encode($data));
-        echo '</pre>';
+        $configInstancesThatCanBeNotFound = [
+            Route::class => ApiPathEnum::ROUTES,
+            Listener::class => ApiPathEnum::LISTENERS,
+            AbstractApplication::class => ApiPathEnum::APPLICATIONS,
+            Upstream::class => ApiPathEnum::UPSTREAMS
+        ];
 
         try {
-            $request->setMethod(HttpMethodsEnum::PUT->value)->send(
+            $request->setMethod(HttpMethodsEnum::PUT)->send(
                 $this->getApiEndpoint(),
                 true,
                 ['json' => $data]
             );
         } catch (UnitException $e) {
+            foreach ($configInstancesThatCanBeNotFound as $configInstance => $path) {
+                if ($this instanceof $configInstance) {
+                    $request->setMethod(HttpMethodsEnum::PUT)->send(
+                        $path->value,
+                        true,
+                        ['json' => (object)[]]
+                    );
+
+                    $request->setMethod(HttpMethodsEnum::PUT)->send(
+                        $this->getApiEndpoint(),
+                        true,
+                        ['json' => $data]
+                    );
+                    return;
+                }
+            }
             throw new UnitException($e->getMessage());
         }
     }
@@ -72,7 +95,7 @@ trait CanUpload
     public function remove(UnitRequest $request)
     {
         try {
-            $request->setMethod(HttpMethodsEnum::DELETE->value)->send($this->getApiEndpoint());
+            $request->setMethod(HttpMethodsEnum::DELETE)->send($this->getApiEndpoint());
         } catch (UnitException $e) {
             throw new UnitException($e->getMessage());
         }